requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
python-dotenv==1.0.0
openai==0.28.1
requests==2.31.0
python-multipart==0.0.6
uuid==1.30
slowapi==0.1.8
redis==5.0.1
httpx==0.25.2
